docker_image/requirements.txt
Flask
cassandra-driver
tensorflow
pillow